+++ src/bot/handler/everyoneHandler.py
@@ -4,31 +4,29 @@ from telegram.update import Update
 
 from bot.handler.abstractHandler import AbstractHandler
 from bot.message.replier import Replier
-from config.contents import no_groups
+from config.contents import mention_failed
 from exception.invalidActionException import InvalidActionException
 from exception.notFoundException import NotFoundException
 from repository.chatRepository import ChatRepository
+from repository.userRepository import UserRepository
 from utils.messageBuilder import MessageBuilder
 
 
-class GroupsHandler(AbstractHandler):
+class EveryoneHandler(AbstractHandler):
     bot_handler: CommandHandler
     chat_repository: ChatRepository
-    action: str = 'groups'
+    user_repository: UserRepository
+    action: str = 'everyone'
 
     def __init__(self) -> None:
         self.bot_handler = CommandHandler(self.action, self.wrap)
         self.chat_repository = ChatRepository()
+        self.user_repository = UserRepository()
 
     def handle(self, update: Update, context: CallbackContext) -> None:
         try:
-            chat = self.chat_repository.get(self.inbound.chat_id)
-            if not chat.groups:
-                raise NotFoundException
+            users = self.chat_repository.get_users_for_group(self.inbound.chat_id, self.inbound.group_name)
 
-            Replier.html(update, MessageBuilder.group_message(chat.groups))
-        except NotFoundException:
-            raise InvalidActionException(no_groups)
-
-    def is_group_specific(self) -> bool:
-        return False
+            Replier.markdown(update, MessageBuilder.mention_message(users))
+        except NotFoundException as e:
+            raise InvalidActionException(mention_failed) from e