+++ src/bot/handler/startHandler.py
@@ -1,25 +1,35 @@
-from config.contents import start_text
-from logger import Logger
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.ext.callbackcontext import CallbackContext
 from telegram.ext.commandhandler import CommandHandler
 from telegram.update import Update
 
 from bot.handler.abstractHandler import AbstractHandler
-from bot.message.messageData import MessageData
+from bot.message.replier import Replier
+from config.contents import start_text
+from logger import Logger
 
 
 class StartHandler(AbstractHandler):
     bot_handler: CommandHandler
+    action: str = 'start'
 
     def __init__(self) -> None:
-        self.bot_handler = CommandHandler('start', self.handle)
+        self.bot_handler = CommandHandler(self.action, self.wrap)
 
     def handle(self, update: Update, context: CallbackContext) -> None:
-        self.reply_markdown(update, start_text)
-        self.log_action(MessageData.create_from_arguments(update, context))
+        markup = InlineKeyboardMarkup(
+            [
+                [
+                    InlineKeyboardButton('Inline Mode', switch_inline_query_current_chat='example'),
+                ],
+                [
+                    InlineKeyboardButton('GitHub', url='https://github.com/miloszowi/everyone-mention-telegram-bot'),
+                    InlineKeyboardButton('Creator', url='https://t.me/miloszowi')
+                ]
+            ]
+        )
 
-    def get_bot_handler(self) -> CommandHandler:
-        return self.bot_handler
+        Replier.html(update, start_text, markup)
 
-    def log_action(self, message_data: MessageData) -> None:
-        Logger.info(f'User {message_data.username} called /start for {message_data.chat_id}')
+    def is_group_specific(self) -> bool:
+        return False