+++ src/exception/invalidActionException.py
@@ -1,0 +1,2 @@
+class InvalidActionException(Exception):
+    pass