+++ src/logger.py
@@ -3,7 +3,10 @@ from __future__ import annotations
 import logging
 import os
 
+from bot.message.inboundMessage import InboundMessage
 
+
+# noinspection SpellCheckingInspection
 class Logger:
     action_logger: str = 'action-logger'
     action_logger_file: str = '/var/log/bot/action.log'
@@ -11,9 +14,7 @@ class Logger:
     main_logger: str = 'main-logger'
     main_logger_file: str = '/var/log/bot/app.log'
 
-    formatter: str = logging.Formatter('%(asctime)s[%(levelname)s]: %(message)s')
-
-    def setup(self) -> None:
+    def __init__(self):
         self.configure(self.action_logger, self.action_logger_file, logging.INFO)
         self.configure(self.main_logger, self.main_logger_file, logging.ERROR)
 
@@ -23,21 +24,38 @@ class Logger:
             os.makedirs(directory)
 
         logger = logging.getLogger(logger_name)
+        logger.propagate = False
+
         file_handler = logging.FileHandler(log_file, mode='w')
-        file_handler.setFormatter(self.formatter)
+        formatter = logging.Formatter('%(asctime)s [%(levelname)s]: %(message)s', datefmt='%H:%M:%S %Y/%m/%d')
+        file_handler.setFormatter(formatter)
         stream_handler = logging.StreamHandler()
-        stream_handler.setFormatter(self.formatter)
+        stream_handler.setFormatter(formatter)
 
         logger.setLevel(level)
         logger.addHandler(file_handler)
         logger.addHandler(stream_handler)
 
     @staticmethod
-    def get_logger(logger_name) -> logging.Logger:
+    def register() -> None:
+        Logger()
+
+    @staticmethod
+    def get(logger_name: str) -> logging.Logger:
         return logging.getLogger(logger_name)
 
+    @staticmethod
     def info(message: str) -> None:
-        Logger.get_logger(Logger.action_logger).info(message)
+        Logger.get(Logger.action_logger).info(message)
 
+    @staticmethod
     def error(message: str) -> None:
-        Logger.get_logger(Logger.main_logger).error(message)
+        Logger.get(Logger.main_logger).error(message)
+
+    @staticmethod
+    def exception(exception: Exception) -> None:
+        Logger.get(Logger.main_logger).exception(exception)
+
+    @staticmethod
+    def action(inbound: InboundMessage, action: str) -> None:
+        Logger.info(f'User {inbound.username}({inbound.user_id}) called {action.upper()} for {inbound.chat_id}({inbound.group_name})')